backend/requirements.txt
# Production requirements - excludes testing dependencies
# Use this for production deployments

Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.3

# User Management System
django-extensions==3.2.3
django-filter==23.5
django-ratelimit==4.1.0
django-activity-stream==2.0.0
django-auditlog==3.2.1
drf-spectacular==0.26.5
Pillow==10.1.0
posthog==3.5.0
pyotp==2.9.0
qrcode[pil]==7.4.2

# Environment variable management
django-environ==0.12.0
python-dotenv==1.0.0

# Email providers and AWS SDK
django-anymail==10.3
boto3==1.35.32

# Payment providers
stripe==14.0.1

# Production Server
gunicorn==21.2.0

# Celery for background tasks
celery==5.3.4
redis==5.0.1

# Security (optional - for CSP support)
# django-csp==3.8

# Database Management
cryptography>=41.0.0
# Optional: For MySQL support (uncomment if needed)
# mysqlclient>=2.1.0
# Note: psycopg2-binary is already included above for PostgreSQL support

anthropic==0.25.0
